+++ doc/changelog.rst
@@ -11,6 +11,7 @@ Unreleased
 ==========
 
 * Fixed: ``--regex-type`` to accept ``pcre`` instead of ``posix``. Regular expressions always use Python's regex library. Posix regex is not supported.
+* Added: Open Graph and Twitter Card element links extraction.
 
 
 1.0 (2015-03-14)